+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.6.0 (unreleased)
 ------------------
 
+- #2524 Allow to edit the result capture date
 - #2527 Fix clients groups are displayed on labcontact's user creation form
 - #2523 Fix getting ResultsRange field exception
 - #2526 Highlight edit tabs with errors in red